refactor: RowMapper를 함수형 인터페이스로 분리
